Remove old login activity
