+++ app/src/main/java/de/synyx/android/meeroo/MeerooApplication.java
@@ -6,7 +6,7 @@ import de.synyx.android.meeroo.business.calendar.CalendarModeService;
 import de.synyx.android.meeroo.config.MainConfig;
 import de.synyx.android.meeroo.config.Registry;
 import de.synyx.android.meeroo.preferences.PreferencesService;
-import de.synyx.android.meeroo.screen.login.mvvm.MVVMLoginViewModelFactory;
+import de.synyx.android.meeroo.screen.login.LoginViewModelFactory;
 import de.synyx.android.meeroo.util.proxy.PermissionManager;
 import de.synyx.android.meeroo.util.proxy.ProxyConfig;
 
@@ -25,9 +25,9 @@ public class MeerooApplication extends Application {
         initConfig();
 
         // for login
-        Registry.put(MVVMLoginViewModelFactory.class,
-            new MVVMLoginViewModelFactory(Registry.get(PreferencesService.class),
-                Registry.get(PermissionManager.class), Registry.get(CalendarModeService.class)));
+        Registry.put(LoginViewModelFactory.class,
+            new LoginViewModelFactory(Registry.get(PreferencesService.class), Registry.get(PermissionManager.class),
+                Registry.get(CalendarModeService.class)));
     }
