+++ app/src/main/java/de/synyx/android/meeroo/screen/login/LoginFragment.java
@@ -1,46 +1,53 @@
 package de.synyx.android.meeroo.screen.login;
 
-import android.Manifest;
-
-import android.app.AlertDialog;
-
-import android.content.DialogInterface;
 import android.content.Intent;
 
 import android.content.pm.PackageManager;
 
-import android.net.Uri;
-
 import android.os.Bundle;
-
-import android.provider.Settings;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import android.widget.Button;
 import android.widget.ProgressBar;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import androidx.fragment.app.Fragment;
 
+import androidx.lifecycle.ViewModelProviders;
+
 import de.synyx.android.meeroo.R;
 import de.synyx.android.meeroo.config.Registry;
-import de.synyx.android.meeroo.util.proxy.PermissionManager;
+import de.synyx.android.meeroo.domain.CalendarMode;
 
-import java.util.List;
+import static android.app.Activity.RESULT_OK;
 
 
 /**
  * @author  Julian Heetel - heetel@synyx.de
  */
-public class LoginFragment extends Fragment implements LoginContract.LoginView {
+public class LoginFragment extends Fragment {
 
-    private static final int REQUEST_PERMISSION_SETTING = 1;
-    private LoginContract.LoginPresenter presenter;
+    static final int PERMISSION_REQUEST_CODE = 1602;
+    static final int REQUEST_ACCOUNT = 139;
+    private LoginViewModel viewModel;
+
+    private TextView step2;
+    private TextView stepAccount;
+    private TextView step3;
+    private TextView stepMode;
     private ProgressBar progressBar;
+    private TextView errorText;
+    private Button errorExitButton;
+    private Button errorRetryButton;
+    private TextView modeText;
+    private Button modeResourcesButton;
+    private Button modeCalendarButton;
 
     @Nullable
     @Override
@@ -49,18 +56,117 @@ public class LoginFragment extends Fragment implements LoginContract.LoginView {
 
         View view = inflater.inflate(R.layout.fragment_login, container, false);
 
+        step2 = view.findViewById(R.id.step_2);
+        stepAccount = view.findViewById(R.id.step_account);
+        step3 = view.findViewById(R.id.step_3);
+        stepMode = view.findViewById(R.id.step_mode);
+
         progressBar = view.findViewById(R.id.login_progress_bar);
+
+        errorText = view.findViewById(R.id.login_error_text);
+        errorExitButton = view.findViewById(R.id.login_error_exit_button);
+        errorRetryButton = view.findViewById(R.id.login_error_retry_button);
+        errorExitButton.setOnClickListener(this::onExitClicked);
+        errorRetryButton.setOnClickListener(this::onRetryClicked);
+
+        modeText = view.findViewById(R.id.login_select_mode_text);
+        modeResourcesButton = view.findViewById(R.id.login_mode_resources_button);
+        modeCalendarButton = view.findViewById(R.id.login_mode_calendar_button);
+        modeResourcesButton.setOnClickListener(this::onModeResourcesClicked);
+        modeCalendarButton.setOnClickListener(this::onModeCalendarClicked);
 
         return view;
     }
 
 
     @Override
-    public void onStart() {
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
 
-        super.onStart();
+        super.onViewCreated(view, savedInstanceState);
 
-        presenter.start();
+        viewModel = ViewModelProviders.of(this, Registry.get(LoginViewModelFactory.class)).get(LoginViewModel.class);
+
+        viewModel.getLoginStep().observe(this, this::onLoginStepChanged);
+    }
+
+
+    private void onRetryClicked(View view) {
+
+        displayProgress(true);
+        displayErrorRetry(false);
+
+        // post same value to retry step
+        viewModel.getLoginStep().postValue(viewModel.getLoginStep().getValue());
+    }
+
+
+    private void onExitClicked(View view) {
+
+        if (getActivity() != null) {
+            getActivity().finish();
+        }
+    }
+
+
+    private void onModeResourcesClicked(View view) {
+
+        viewModel.saveCalendarMode(CalendarMode.RESOURCES);
+    }
+
+
+    private void onModeCalendarClicked(View view) {
+
+        viewModel.saveCalendarMode(CalendarMode.CALENDAR);
+    }
+
+
+    private void onLoginStepChanged(LoginStep loginStep) {
+
+        if (LoginStep.PERMISSIONS == loginStep) {
+            enableStepAccount(false);
+            enableStepMode(false);
+            viewModel.askForPermissions(this);
+        } else if (LoginStep.ACCOUNT == loginStep) {
+            enableStepAccount(true);
+            enableStepMode(false);
+            viewModel.stepAccount(this);
+        } else if (LoginStep.MODE == loginStep) {
+            enableStepAccount(true);
+            enableStepMode(true);
+            displayProgress(false);
+            viewModel.stepMode();
+            displayModeSelection();
+        } else if (LoginStep.FINISHED == loginStep) {
+            onLoginFinished();
+        }
+    }
+
+
+    private void displayModeSelection() {
+
+        modeText.setVisibility(View.VISIBLE);
+        modeResourcesButton.setVisibility(View.VISIBLE);
+        modeCalendarButton.setVisibility(View.VISIBLE);
+    }
+
+
+    private void displayProgress(boolean visible) {
+
+        progressBar.setVisibility(visible ? View.VISIBLE : View.GONE);
+    }
+
+
+    private void enableStepAccount(boolean enabled) {
+
+        step2.setEnabled(enabled);
+        stepAccount.setEnabled(enabled);
+    }
+
+
+    private void enableStepMode(boolean enabled) {
+
+        step3.setEnabled(enabled);
+        stepMode.setEnabled(enabled);
     }
 
 
@@ -68,163 +174,66 @@ public class LoginFragment extends Fragment implements LoginContract.LoginView {
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
         @NonNull int[] grantResults) {
 
-        if (requestCode == REQUEST_PERMISSION_SETTING) {
+        if (requestCode == PERMISSION_REQUEST_CODE) {
             if (permissions.length == 0) {
                 // the dialog was cancelled, e.g. orientation change
                 return;
             }
 
             for (int i = 0; i < permissions.length; i++) {
-                if (grantResults[i] == PackageManager.PERMISSION_DENIED) {
-                    onPermissionDenied(permissions[i]);
+                if (PackageManager.PERMISSION_DENIED == grantResults[i]) {
+                    onPermissionDenied();
 
                     return;
                 }
             }
 
-            presenter.onAllPermissionsGranted();
+            onAllPermissionsGranted();
         }
+    }
+
+
+    private void onPermissionDenied() {
+
+        displayProgress(false);
+        displayErrorRetry(true);
+    }
+
+
+    private void displayErrorRetry(boolean visible) {
+
+        errorText.setVisibility(visible ? View.VISIBLE : View.GONE);
+        errorExitButton.setVisibility(visible ? View.VISIBLE : View.GONE);
+        errorRetryButton.setVisibility(visible ? View.VISIBLE : View.GONE);
+    }
+
+
+    private void onAllPermissionsGranted() {
+
+        viewModel.getLoginStep().postValue(LoginStep.ACCOUNT);
     }
 
 
     @Override
-    public void askForPermissions(List<String> neededPermissions) {
+    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
 
-        String[] permissionsToRequest = new String[neededPermissions.size()];
+        if (requestCode == REQUEST_ACCOUNT && resultCode == RESULT_OK) {
+            String accountName = data.getStringExtra("authAccount");
 
-        Registry.get(PermissionManager.class)
-            .requestPermission(this, neededPermissions.toArray(permissionsToRequest), REQUEST_PERMISSION_SETTING);
-    }
-
-
-    @Override
-    public void showAccountSelection(String[] accounts) {
-
-        new AlertDialog.Builder(getActivity()).setTitle(R.string.selectAccount)
-            .setItems(accounts, (dialog, which) -> presenter.onAccountSelected(accounts[which]))
-            .setCancelable(false).create().show();
-    }
-
-
-    @Override
-    public void showCalendarModeSelection(String[] calenderModes) {
-
-        new AlertDialog.Builder(getActivity()).setTitle(R.string.selectMode)
-            .setItems(calenderModes, (dialog, which) -> presenter.onCalendarModeSelected(calenderModes[which]))
-            .setCancelable(false).create().show();
-    }
-
-
-    @Override
-    public void showErrorDialog() {
-
-        new AlertDialog.Builder(getActivity()).setMessage(getString(R.string.noCalendarsError))
-            .setTitle(getString(R.string.calendarError))
-            .setPositiveButton(R.string.close, ((dialog, which) -> presenter.onErrorDialogCloseButtonClicked()))
-            .setCancelable(false).create().show();
-    }
-
-
-    @Override
-    public void showProgress() {
-
-        progressBar.setVisibility(View.VISIBLE);
-//        progressText.setVisibility(View.VISIBLE);
-//        progressText.setText(getString(R.string.calendarPending));
-    }
-
-
-    public void setPresenter(LoginContract.LoginPresenter presenter) {
-
-        this.presenter = presenter;
-    }
-
-
-    private void onPermissionDenied(String permission) {
-
-        if (Registry.get(PermissionManager.class).shouldShowRequestPermissionRationale(this, permission)) {
-            showPermissionNeededDialog(permission,
-                (dialog, which) -> handlePermissionsNeededRationaleDialogResult(which));
-        } else {
-            showPermissionNeededDialog(permission, (dialog, which) -> handlePermissionsNeededDialogResult(which));
-        }
-    }
-
-
-    private void handlePermissionsNeededDialogResult(int which) {
-
-        if (which == DialogInterface.BUTTON_POSITIVE) {
-            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-            Uri uri = Uri.fromParts("package", getActivity().getPackageName(), null);
-            intent.setData(uri);
-            startActivityForResult(intent, REQUEST_PERMISSION_SETTING);
-        } else if (which == DialogInterface.BUTTON_NEGATIVE) {
-            presenter.onAllPermissionsGranted();
-        }
-    }
-
-
-    private void handlePermissionsNeededRationaleDialogResult(int which) {
-
-        if (which == DialogInterface.BUTTON_POSITIVE) {
-            presenter.askForPermissionsAgain();
-        } else if (which == DialogInterface.BUTTON_NEGATIVE) {
-            presenter.onAllPermissionsGranted();
-        }
-    }
-
-
-    private void showPermissionNeededDialog(String permission, DialogInterface.OnClickListener clickListener) {
-
-        if (getActivity() == null) {
-            return;
+            viewModel.saveAccount(accountName);
+        } else if (requestCode == REQUEST_ACCOUNT) {
+            displayProgress(false);
+            displayErrorRetry(true);
         }
 
-        switch (permission) {
-            case Manifest.permission.READ_CALENDAR:
-                showChoiceDialog(getString(R.string.dialog_calendar_permission_massage),
-                    getString(R.string.dialog_calendar_permission_title),
-                    getString(R.string.permission_dialog_positive), getString(R.string.permission_dialog_negative),
-                    clickListener);
-                break;
-
-            case Manifest.permission.WRITE_CALENDAR:
-                showChoiceDialog(getString(R.string.dialog_calendar_permission_massage),
-                    getString(R.string.dialog_calendar_permission_title),
-                    getString(R.string.permission_dialog_positive), getString(R.string.permission_dialog_negative),
-                    clickListener);
-                break;
-
-            case Manifest.permission.GET_ACCOUNTS:
-                showChoiceDialog(getString(R.string.dialog_accounts_permission_massage),
-                    getString(R.string.dialog_accounts_permission_title),
-                    getString(R.string.permission_dialog_positive), getString(R.string.permission_dialog_negative),
-                    clickListener);
-                break;
-
-            case Manifest.permission.READ_SYNC_SETTINGS:
-                showChoiceDialog(getString(R.string.dialog_sync_settings_permission_massage),
-                    getString(R.string.dialog_sync_settings_permission_title),
-                    getString(R.string.permission_dialog_positive), getString(R.string.permission_dialog_negative),
-                    clickListener);
-                break;
-
-            case Manifest.permission.READ_CONTACTS:
-                showChoiceDialog(getString(R.string.dialog_contacts_permission_massage),
-                    getString(R.string.dialog_contacts_permission_title),
-                    getString(R.string.permission_dialog_positive), getString(R.string.permission_dialog_negative),
-                    clickListener);
-                break;
-        }
+        super.onActivityResult(requestCode, resultCode, data);
     }
 
 
-    public void showChoiceDialog(String message, String title, String positiveButtonText, String negativeButtonText,
-        final DialogInterface.OnClickListener clickListener) {
+    private void onLoginFinished() {
 
-        new AlertDialog.Builder(getActivity()).setTitle(title)
-            .setMessage(message)
-            .setPositiveButton(positiveButtonText, clickListener)
-            .setNegativeButton(negativeButtonText, clickListener).setCancelable(false).create().show();
+        if (getActivity() != null) {
+            ((LoginActivity) getActivity()).startMainActivity();
+        }
     }
 }
