--- app/src/main/java/de/synyx/android/meeroo/screen/login/LoginListener.java
@@ -1,15 +1,0 @@
-package de.synyx.android.meeroo.screen.login;
-
-import de.synyx.android.meeroo.domain.CalendarMode;
-
-
-/**
- * @author  Julia Dasch - dasch@synyx.de
- */
-public interface LoginListener {
-
-    void onErrorCloseButtonClick();
-
-
-    void onCalenderModeClick(CalendarMode calendarMode);
-}
