+++ app/src/main/java/de/synyx/android/meeroo/screen/login/LoginStep.java
@@ -1,4 +1,4 @@
-package de.synyx.android.meeroo.screen.login.mvvm;
+package de.synyx.android.meeroo.screen.login;
 
 /**
  * @author  Julian Heetel - heetel@synyx.de
